+++ Modules/Microsoft365DSC/Modules/M365DSCReverse.psm1
@@ -197,6 +197,16 @@ function Start-M365DSCConfigurationExtract
             $ComponentsToSkip += $ExcludeComponents
         }
 
+        $resourcesInBothIncludeAndExclude = Compare-Object -ReferenceObject $Components `
+            -DifferenceObject $ComponentsToSkip -ExcludeDifferent -IncludeEqual
+        if ($resourcesInBothIncludeAndExclude.Count -gt 0)
+        {
+            foreach ($resource in $resourcesInBothIncludeAndExclude)
+            {
+                Write-Warning -Message "The component '$($resource.InputObject)' was specified in both -Components and -ExcludeComponents parameters. It will be excluded from the export."
+            }
+        }
+
         # Check to validate that based on the received authentication parameters
         # we are allowed to export the selected components.
         $AuthMethods = @()
@@ -757,10 +767,12 @@ function Start-M365DSCConfigurationExtract
                 # Check for ErrorAction Preference
                 $parameters.Add('ErrorAction', $using:ErrorActionPreference)
                 $Global:M365DSCExportResourceTypes += $resourceName
+
                 try
                 {
                     $exportOutput = Export-TargetResource @parameters
                     $exportString.Append($exportOutput) | Out-Null
+                    ($using:synchronizedHashtable).ResourcesResult.Add($resourceName, $exportString.ToString())
                 }
                 catch
                 {
@@ -771,7 +783,6 @@ function Start-M365DSCConfigurationExtract
                     }
                 }
             }
-            ($using:synchronizedHashtable).ResourcesResult.Add($resourceName, $exportString.ToString())
         }
 
         if ($Parallel)